show_asset fixed (still need to switch to fopen)

+++ user.h
@@ -87,7 +87,7 @@ void handle_list_response(char *status, char *buffer, char *auction_list);
 void list(char *ASIP, char *ASport);
 
 /* Transfers data from the socket to a local file */
-void copy_from_socket_to_file(int written, int size, int fd, struct addrinfo *res, int file_fd);
+void copy_from_socket_to_file(long written, int size, int fd, struct addrinfo *res, int file_fd);
 
 /* Handles show_asset server responses */
 void handle_show_asset_response(char *status, char *prefix, int fd, struct addrinfo *res);
